backend/ai/requirements.txt
deepgram-sdk
pyaudio
python-dotenv
requests
openai
cerebras-cloud-sdk
llama-index
llama-index-embeddings-openai